requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
crontab==1.0.1
Django==4.2.6
django-apscheduler==0.6.2
django-cron==0.6.0
django-crontab==0.7.1
djangorestframework==3.14.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzlocal==5.2
